Closes #66
