+++ RDF_Engine/Query/Code analysis/Reflection/GetCodeName.cs
@@ -22,7 +22,6 @@
 
 using System;
 using System.CodeDom;
-using Microsoft.CodeDom.Providers.DotNetCompilerPlatform;
 using System.Collections;
 using System.Collections.Generic;
 using System.ComponentModel;
@@ -32,6 +31,7 @@ using System.Text;
 using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 using System.IO;
+using Microsoft.CSharp;
 
 namespace BH.Engine.RDF
 {
@@ -41,7 +41,7 @@ namespace BH.Engine.RDF
 		public static string GetCodeName(this Type type, bool includeNamespace = false)
 		{
             string RoslynPath = Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location) + "\\roslyn\\csc.exe";
-            m_CodeDomProvider = new CSharpCodeProvider(new ProviderOptions(RoslynPath, 0));
+            m_CodeDomProvider = new CSharpCodeProvider();
 
             m_CodeDomProvider = m_CodeDomProvider ?? new CSharpCodeProvider(); // cache provider.
