+++ RDF_Engine/Query/Code analysis/Reflection/IsGenericTypeWithBHoMArgs.cs
@@ -20,7 +20,6 @@
  * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
  */
 
-using BH.Adapter;
 using BH.oM.Base;
 using System;
 using System.Collections;
