+++ RDF_Prototypes.sln
@@ -5,8 +5,6 @@ VisualStudioVersion = 17.1.32421.90
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RDF_ConsoleApp", "RDF_ConsoleApp\RDF_ConsoleApp.csproj", "{E2EC7085-D4F5-4C1A-B2F0-FBFCBA0D1938}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RDF_Engine", "RDF_Engine\RDF_Engine.csproj", "{9424302A-AAB4-41EA-B109-157BA549DE96}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RDF_oM", "RDF_oM\RDF_oM.csproj", "{35CB5890-7BC1-4F9C-A63F-771F71124D6A}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{5B85B6B2-63CD-4BB1-B1D6-D8C23B5C20C2}"
@@ -18,32 +16,53 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RDF_Tests", "RDF_Tests\RDF_
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TTL_Adapter", "TTL_Adapter\TTL_Adapter.csproj", "{663143C2-5E21-4E1F-8E06-6E0FFB3DE74D}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TTL_Engine", "TTL_Engine\TTL_Engine.csproj", "{B43D4315-1673-4B88-AD97-FAB0586331C7}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RDF_Engine", "RDF_Engine\RDF_Engine.csproj", "{A25BACE3-D136-4E6B-8BAA-ECEB522F5EE4}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
 		Release|Any CPU = Release|Any CPU
+		Test|Any CPU = Test|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
 		{E2EC7085-D4F5-4C1A-B2F0-FBFCBA0D1938}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{E2EC7085-D4F5-4C1A-B2F0-FBFCBA0D1938}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E2EC7085-D4F5-4C1A-B2F0-FBFCBA0D1938}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{E2EC7085-D4F5-4C1A-B2F0-FBFCBA0D1938}.Release|Any CPU.Build.0 = Release|Any CPU
-		{9424302A-AAB4-41EA-B109-157BA549DE96}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{9424302A-AAB4-41EA-B109-157BA549DE96}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{9424302A-AAB4-41EA-B109-157BA549DE96}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{9424302A-AAB4-41EA-B109-157BA549DE96}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E2EC7085-D4F5-4C1A-B2F0-FBFCBA0D1938}.Test|Any CPU.ActiveCfg = Test|Any CPU
+		{E2EC7085-D4F5-4C1A-B2F0-FBFCBA0D1938}.Test|Any CPU.Build.0 = Test|Any CPU
 		{35CB5890-7BC1-4F9C-A63F-771F71124D6A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{35CB5890-7BC1-4F9C-A63F-771F71124D6A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{35CB5890-7BC1-4F9C-A63F-771F71124D6A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{35CB5890-7BC1-4F9C-A63F-771F71124D6A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{35CB5890-7BC1-4F9C-A63F-771F71124D6A}.Test|Any CPU.ActiveCfg = Test|Any CPU
+		{35CB5890-7BC1-4F9C-A63F-771F71124D6A}.Test|Any CPU.Build.0 = Test|Any CPU
 		{8BD3A97A-E20A-4F23-B48F-5730F28458CF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{8BD3A97A-E20A-4F23-B48F-5730F28458CF}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{8BD3A97A-E20A-4F23-B48F-5730F28458CF}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{8BD3A97A-E20A-4F23-B48F-5730F28458CF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8BD3A97A-E20A-4F23-B48F-5730F28458CF}.Test|Any CPU.ActiveCfg = Test|Any CPU
+		{8BD3A97A-E20A-4F23-B48F-5730F28458CF}.Test|Any CPU.Build.0 = Test|Any CPU
 		{663143C2-5E21-4E1F-8E06-6E0FFB3DE74D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{663143C2-5E21-4E1F-8E06-6E0FFB3DE74D}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{663143C2-5E21-4E1F-8E06-6E0FFB3DE74D}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{663143C2-5E21-4E1F-8E06-6E0FFB3DE74D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{663143C2-5E21-4E1F-8E06-6E0FFB3DE74D}.Test|Any CPU.ActiveCfg = Test|Any CPU
+		{663143C2-5E21-4E1F-8E06-6E0FFB3DE74D}.Test|Any CPU.Build.0 = Test|Any CPU
+		{B43D4315-1673-4B88-AD97-FAB0586331C7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B43D4315-1673-4B88-AD97-FAB0586331C7}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B43D4315-1673-4B88-AD97-FAB0586331C7}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B43D4315-1673-4B88-AD97-FAB0586331C7}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B43D4315-1673-4B88-AD97-FAB0586331C7}.Test|Any CPU.ActiveCfg = Test|Any CPU
+		{B43D4315-1673-4B88-AD97-FAB0586331C7}.Test|Any CPU.Build.0 = Test|Any CPU
+		{A25BACE3-D136-4E6B-8BAA-ECEB522F5EE4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A25BACE3-D136-4E6B-8BAA-ECEB522F5EE4}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A25BACE3-D136-4E6B-8BAA-ECEB522F5EE4}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A25BACE3-D136-4E6B-8BAA-ECEB522F5EE4}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A25BACE3-D136-4E6B-8BAA-ECEB522F5EE4}.Test|Any CPU.ActiveCfg = Test|Any CPU
+		{A25BACE3-D136-4E6B-8BAA-ECEB522F5EE4}.Test|Any CPU.Build.0 = Test|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
