+++ RDF_Tests/FromTTLTests.cs
@@ -37,8 +37,8 @@ using BH.oM.Structure.Elements;
 using NUnit.Framework;
 using BH.Adapters.TTL;
 using Compute = BH.Engine.RDF.Compute;
-using BH.Adapters.TTL;
-using Convert = BH.Adapters.TTL.Convert;
+using Convert = BH.Engine.Adapters.TTL.Convert;
+using BH.Engine.Adapters.TTL;
 
 namespace BH.Test.RDF
 {
@@ -74,7 +74,7 @@ namespace BH.Test.RDF
 
             Assert.IsTTLParsable(TTLGraph); // This MUST return an encoded customData dictionary - its entry is not seen as property!
 
-            var objs = Adapters.TTL.Convert.FromTTL(TTLGraph).Item1;
+            var objs = Engine.Adapters.TTL.Convert.FromTTL(TTLGraph).Item1;
 
             Assert.IsEqual(bhomObj, objs.First());
         }
@@ -123,7 +123,7 @@ namespace BH.Test.RDF
                 }
             };
 
-            string TTLGraph = Adapters.TTL.Compute.TTLGraph(new List<object>() { bhomObject }, m_ontologySettings);
+            string TTLGraph = Engine.Adapters.TTL.Compute.TTLGraph(new List<object>() { bhomObject }, m_ontologySettings);
 
             Assert.IsTTLParsable(TTLGraph);
 
@@ -170,7 +170,7 @@ namespace BH.Test.RDF
 
             Assert.IsTTLParsable(TTLGraph);
 
-            Output<List<object>, OntologySettings> bhomObjects_ontologySettings = BH.Adapters.TTL.Convert.FromTTL(TTLGraph);
+            Output<List<object>, OntologySettings> bhomObjects_ontologySettings = BH.Engine.Adapters.TTL.Convert.FromTTL(TTLGraph);
 
             Assert.IsEqual(bhomObjects_ontologySettings.Item2, customOntologySettings);
         }
@@ -280,7 +280,7 @@ namespace BH.Test.RDF
 
             Assert.IsTTLParsable(TTLGraph);
 
-            Output<List<object>, OntologySettings> convertedObjs = BH.Adapters.TTL.Convert.FromTTL(TTLGraph);
+            Output<List<object>, OntologySettings> convertedObjs = BH.Engine.Adapters.TTL.Convert.FromTTL(TTLGraph);
 
             Assert.IsEqual(customObject1, convertedObjs.Item1.First());
         }
