+++ TTL_Engine/Convert/ToTTL/TTLDataProperties.cs
@@ -31,7 +31,7 @@ using System.Reflection;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace BH.Adapters.TTL
+namespace BH.Engine.Adapters.TTL
 {
     public static partial class Convert
     {
