+++ TTL_Engine/Convert/ToTTL/TTLObjectProperties.cs
@@ -31,40 +31,44 @@ using System.Reflection;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace BH.Adapters.TTL
+namespace BH.Engine.Adapters.TTL
 {
     public static partial class Convert
     {
-        private static List<string> TTLClasses(this CSharpGraph cSharpGraph, LocalRepositorySettings localRepositorySettings)
+        private static List<string> TTLObjectProperties(this CSharpGraph cSharpGraph, LocalRepositorySettings localRepositorySettings)
         {
-            List<string> TTLClasses = new List<string>();
+            List<string> TTLObjectProperties = new List<string>();
 
-            foreach (var classType in cSharpGraph.Classes)
+            for (int i = 0; i < cSharpGraph.ObjectProperties.Count; i++)
             {
-                string TTLClass = "";
+                var rel = cSharpGraph.ObjectProperties.ElementAt(i);
 
-                // Declaration with Uri
-                string typeUri = classType.OntologyUri(cSharpGraph.OntologySettings.TBoxSettings, localRepositorySettings).ToString();
-                TTLClass += $"### {typeUri}";
-
-                // Class Identifier
-                TTLClass += $"\n:{classType.UniqueNodeId()} rdf:type owl:Class;";
-
-                // Subclasses
-                List<Type> parentTypes = classType.BaseTypesNoRedundancy().Where(t => t.IsOntologyClass(cSharpGraph.OntologySettings.TBoxSettings)).ToList();
-
-                foreach (Type subClass in parentTypes)
+                try
                 {
-                    TTLClass += $"\n\t\trdfs:subClassOf :{subClass.UniqueNodeId()};";
+                    string TTLObjectProperty = "";
+
+                    if (rel.RangeClass == null || rel.PropertyInfo == null)
+                        continue;
+
+                    if (localRepositorySettings.TryComputeURLFromFilePaths)
+                    {
+                        string propertyURI = rel.PropertyInfo.OntologyURI(cSharpGraph.OntologySettings.TBoxSettings, localRepositorySettings).ToString();
+                        TTLObjectProperty += $"\n### {propertyURI}";
+                    }
+                    TTLObjectProperty += $"\n:{rel.PropertyInfo.UniqueNodeId()} rdf:type owl:ObjectProperty ;";
+                    TTLObjectProperty += $"\nrdfs:domain :{rel.DomainClass.UniqueNodeId()} ;";
+                    TTLObjectProperty += $"\nrdfs:range :{rel.RangeClass.UniqueNodeId()} ;";
+                    TTLObjectProperty += "\n" + $@"rdfs:label ""{rel.PropertyInfo.DescriptiveName()}""@en .";
+
+                    TTLObjectProperties.Add(TTLObjectProperty);
                 }
-
-                // Class label
-                TTLClass += "\n\t\t" + $@"rdfs:label ""{classType.DescriptiveName()}""@en .";
-
-                TTLClasses.Add(TTLClass);
+                catch (Exception e)
+                {
+                    Log.RecordError($"Could not add the {nameof(CSharpGraph)}.{nameof(CSharpGraph.ObjectProperties)} at position {i}. Error:\n\t{e.ToString()}");
+                }
             }
 
-            return TTLClasses;
+            return TTLObjectProperties;
         }
     }
 }
